refactored boolRouter and added boolController when the request logic has now been placed to further modularize the code-base.
